Split AudReader from AudLoader and move it to FileFormats
We already separated loading from file format reading for sprites and videos.
