+++ OpenRA.Mods.Common/FileFormats/VqaReader.cs
@@ -12,7 +12,6 @@
 using System;
 using System.IO;
 using OpenRA.FileFormats;
-using OpenRA.Mods.Common.AudioLoaders;
 
 namespace OpenRA.Mods.Common.FileFormats
 {
